// frontend/src/types/HousingData.ts
export interface HousingData {
    id: number;
    censusArea: string;
    totalStarts: number;
    totalComplete: number;
    month: number;
    year: number;
    singleStarts: number;
    semisStarts: number;
    rowStarts: number;
    apartmentStarts: number;
    singlesComplete: number;
    semisComplete: number;
    rowComplete: number;
    apartmentComplete: number;
}